add stdin.fileno() workaround for maya
Autodesk have replaced the stdin object with another `file-like` object
which they use to redirect output to the maya log. This object does not
implement the fileno method which causes rez to crash. In cases where
the method is missing we use the private version of __stdin__ which still
works.